ATLAS-2243: Code improvement suggested by Static code analysis

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v1/EntityGraphRetriever.java
@@ -776,13 +776,13 @@ public final class EntityGraphRetriever {
         relationship.setCreateTime(new Date(GraphHelper.getCreatedTime(edge)));
         relationship.setUpdateTime(new Date(GraphHelper.getModifiedTime(edge)));
 
-        Integer version = GraphHelper.getVersion(edge).intValue();
+        Long version = GraphHelper.getVersion(edge);
 
         if (version == null) {
-            version = Integer.valueOf(1);
+            version = Long.valueOf(1L);
         }
 
-        relationship.setVersion(version.longValue());
+        relationship.setVersion(version);
         relationship.setStatus(GraphHelper.getEdgeStatus(edge));
 
         AtlasVertex end1Vertex = edge.getOutVertex();
